+++ R/FolioBuy.R
@@ -1,23 +1,29 @@
 #' @rdname FolioSell
 #' @export
 
-FolioBuy<- function(loanId, orderId, noteId, Price= NULL, LC_CRED=NULL, quiet=T){
+FolioBuy<- function(loanId=NULL, orderId=NULL, noteId, Price= NULL, LC_CRED=NULL, quiet=T){
+    
+    #check for deprecated variables
+    calls <- names(sapply(match.call(), deparse))[-1]
+    if(any("loanId" %in% calls)) warning("The loanId variable is deprecated")
+    if(any("orderId" %in% calls)) warning("The orderId variable is deprecated")
+    
     LC_CRED<-CheckCred(LC_CRED)
 
-    postURL<- MakeURL(LC_CRED$investorID, "trades/buy")
+    postURL<- paste0("https://api.lendingclub.com/api/investor/v1/secondarymarket/accounts/",
+                     LC_CRED[[1]],
+                     "/orders")
 
-    params<- list("aid"=LC_CRED$investorID,
-                  "notes"= data.frame(
-                      "loanId"= loanId,
-                      "orderId"= orderId,
+    params<-  data.frame(
                       "noteId"= noteId,
-                      "bidPrice"= Price,
-                      stringsAsFactors = F))
+                      "price"= Price,
+                      "orderType"= "BUY",
+                      stringsAsFactors = F)
 
 
     resp<- LC_POST(postURL,params, LC_CRED$key)
     
     if (!quiet) {
-        resp
+        data.frame(do.call("cbind", resp[[1]]), stringsAsFactors = F)
     }
 }