+++ R/LendingClub-deprecated.R
@@ -1,9 +1,8 @@
-## LendingClub-deprecated.r
-#' @title Deprecated functions in package \pkg{LendingClub}.
-#' @description The functions listed below are deprecated and will be defunct in
-#'   the near future. When possible, alternative functions with similar
-#'   functionality are also mentioned. Help pages for deprecated functions are
-#'   available at \code{help("-deprecated")}.
+#' Deprecated functions in package \pkg{LendingClub}.
+#' 
+#' The functions listed below are deprecated and will be removed (defunct) in
+#' the near future. 
+#' - \code{ \link{AccountSummary}}, use \code{ \link{AccountSummary_}}
+#' 
 #' @name LendingClub-deprecated
-#' @keywords internal
 NULL