+++ R/SecondaryListings.R
@@ -6,8 +6,6 @@
 #'
 SecondaryListing<- function(updatedSince=5){
     
-    
-
     LC_CRED<-CheckCred(LC_CRED)
     
     searchURL<-'https://api.lendingclub.com/api/investor/v1/secondarymarket/listings'
@@ -26,27 +24,5 @@ SecondaryListing<- function(updatedSince=5){
         list(content = parsed,
              response = r),
         class = c("folio_api","LendingClub_api"))
-    
-    # # ----    
-    # parsed<- jsonlite::fromJSON(httr::content(r, "text", encoding="UTF-8"),
-    #                             simplifyVector = TRUE)
-    # 
-    # result<- httr::content(result,
-    #                        as="text",
-    #                        type="text/csv",
-    #                        encoding ='UTF-8')
-    # 
-    # parsed<-unlist(strsplit(result, "\n"))
-    # parsed<-lapply(parsed, function(x) unlist(strsplit(x, ",")))
-    # parsed<-matrix(unlist(parsed), ncol=21, nrow=length(parsed), byrow = T)
-    # parsed<- as.data.frame(gsub("\"" ,"", parsed))
-    # colnames(parsed)<- unname(unlist(parsed[1, ]))
-    # 
-    # # result<- LC_GET(searchURL, LC_CRED$key, margin=T)
-    # 
-    # 
-    # class(parsed)<- "SecondaryListing"
-    # 
-    # parsed
     
     }