+++ NEWS.md
@@ -1,3 +1,14 @@
+LendingClub v1.0.5 (Release date: 2018-01-29)
+==============
+
+Changes:  
+
+- Lending Club updated the folio secondary market API.  Several arguments in the FolioBuy and FolioSell 
+functions are no longer required and are deprecated.
+- FolioListing function now able to filter the esults by the time elapsed since the note was 
+updated. The function executes much faster.
+- Fixed typos in the Dashboard output.
+
 LendingClub v1.0.4 (Release date: 2018-01-29)
 ==============
 
@@ -9,5 +20,5 @@ Changes:
 - Deprecated AccountSummary()
 
 
-LendingClub v1.0.3 (Release date: 2018-02-20)
+LendingClub v1.0.3 (Release date: 2017-06-04)
 ==============